Error handlining

+++ utils/readmeToHtml.py
@@ -24,36 +24,38 @@ def rstToHtml(input_file, output_file):
                  writer_name='html')
 
 def main(input_file):
+
+    if os.path.isfile(input_file):
    
-    try:
-        aux_file = f'{os.path.splitext(input_file)[0]}.html'
+        try:
+            aux_file = f'{os.path.splitext(input_file)[0]}.html'
 
-        ext      = os.path.splitext(input_file)[1].split('.')[-1]
-        # Modify the output folder
-        out_file = os.path.join(os.path.dirname(input_file), os.path.basename(aux_file))
-        
-        if ext == 'rst':
-            rstToHtml(input_file, aux_file)
-        elif ext == 'md':
-            mdToHtml(input_file, aux_file)
-        elif ext == 'ipynb':
-            ipynbToHtml(input_file, aux_file)
+            ext      = os.path.splitext(input_file)[1].split('.')[-1]
+            # Modify the output folder
+            out_file = os.path.join(os.path.dirname(input_file), os.path.basename(aux_file))
+            
+            if ext == 'rst':
+                rstToHtml(input_file, aux_file)
+            elif ext == 'md':
+                mdToHtml(input_file, aux_file)
+            elif ext == 'ipynb':
+                ipynbToHtml(input_file, aux_file)
 
-        # Create the final directory if not exists
-        if os.path.isdir(os.path.dirname(out_file)) == False:
-            os.makedirs(os.path.dirname(out_file))
+            # Create the final directory if not exists
+            if os.path.isdir(os.path.dirname(out_file)) == False:
+                os.makedirs(os.path.dirname(out_file))
 
-        # Generate the config file if not exits
-        if os.path.isfile(os.path.join(os.path.dirname(out_file), 'config.json')) == False:
-            with open(os.path.join(os.path.dirname(out_file), 'config.json'), 'w') as file:
-                json.dump({}, file)
+            # Generate the config file if not exits
+            if os.path.isfile(os.path.join(os.path.dirname(out_file), 'config.json')) == False:
+                with open(os.path.join(os.path.dirname(out_file), 'config.json'), 'w') as file:
+                    json.dump({}, file)
 
-        shutil.move(aux_file, out_file)
+            shutil.move(aux_file, out_file)
 
-        print(out_file)
-    except Exception as e:
-        print('Error ->',e)
-        
+            print(out_file)
+        except Exception as e:
+            print('Error ->',e)
+            
 
 #######################################
 ######### TO EXECUTE THE CODE #########
